+++ src/openassetio-python/cmodule/src/_openassetio.hpp
@@ -114,3 +114,6 @@ void registerCppPluginSystemPlugin(const py::module& mod);
 
 // Register the C++ plugin system
 void registerCppPluginSystem(const py::module& mod);
+
+// Register the C++ plugin system manager factory
+void registerCppPluginSystemManagerImplementationFactory(const py::module& mod);